+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.3.0] - 2022-10-11
+
+### Changed
+
+* `Slice` implementation
+* `beginParse()` Cell method
+
 ## [0.2.0] - 2022-08-02
 
 ### Changed